Refactor: change architecture of folders. Add colors and typography to stories.
