+++ src/theme.tsx
@@ -5,6 +5,15 @@ export default {
     white: "#fff",
     blue: "#007ce0",
     navy: "#004175",
+    blue100: "#EBF8FF",
+    blue200: "#BEE3F8",
+    blue300: "#90CDF4",
+    blue400: "#63B3ED",
+    blue500: "#4299E1",
+    blue600: "#3182CE",
+    blue700: "#2B6CB0",
+    blue800: "#2C5282",
+    blue900: "#2A4365",
   },
   space: [0, 4, 8, 16, 32, 64, 128, 256, 512],
 };
