+++ stories/mdx/button.stories.mdx
@@ -1,5 +1,5 @@
-import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
-import {Button} from '../../src/Button.tsx';
+import { Meta, Story, Preview } from "@storybook/addon-docs/blocks";
+import { Button } from "../../src/components/Button.tsx";
 
 <Meta title="MDX|Button" component={Button} />
