+++ stories/mdx/typography.stories.mdx
@@ -1,0 +1,25 @@
+import { Meta, Typeset } from "@storybook/addon-docs/blocks";
+import { typography } from "../../src/styles/typography.tsx";
+
+<Meta title="Design System|Typography" />
+
+# Typography
+
+**Font:** Roboto
+
+**Weights:** 400(regular), 700(bold), 800(extrabold), 900(black)
+
+<Typeset
+  fontSizes={[
+    Number(typography.size.s1),
+    Number(typography.size.s2),
+    Number(typography.size.s3),
+    Number(typography.size.m1),
+    Number(typography.size.m2),
+    Number(typography.size.m3),
+    Number(typography.size.l1),
+    Number(typography.size.l2),
+    Number(typography.size.l3),
+  ]}
+  fontWeight={typography.weight.black}
+/>
